Added: Find Maximum Double value

+++ src/com/maximum/program/MaxProb.java
@@ -6,13 +6,13 @@ public class MaxProb {
 	public void MaxInteg() {
 		// Create the Array For comparing
 		Scanner sc =new Scanner(System.in);
-		int[] arr=new int[3];
+		double[] arr=new double[3];
 		for (int i=0; arr.length > i; i++) {
 			System.out.println("Enter the array value ");
-			arr[i] = sc.nextInt();
+			arr[i] = sc.nextDouble();
 		}
 //    Find The Maximum Program
-		int max=arr[0];
+		double max=arr[0];
 		for (int i=0; i < arr.length; i++) {
 			if (arr[i]> max) {
 				max=arr[i];
